Retried and Cleaned up HTML + CSS
Realized that first version only looked correct on a specific
window size. So started from scratch and redid everything,
making sure to write more neatly and keep the HTML and CSS
in separate files. Also added the buttons on the top right
and at the bottom of the search bar.